+ Added Views (Instance is now replaced).
+ Changed delta to float instead of double.
+ Changed how the mouse button listener works.
+ Added a Texture Creator.
+ Changed EnabledUIComponents to InputEnabledViews.
+ Transform.move() will now return the current Transform.
+ Smaller improvements.

+++ src/se/wiklund/haderengine/ui/style/UIContextMenuStyle.java
@@ -8,11 +8,11 @@ public class UIContextMenuStyle {
     private UIContextMode mode;
     private Texture background;
     private final UIFont font;
-    private float fontSize;
+    private int fontSize;
     private int width;
     private int buttonHeight;
 
-    public UIContextMenuStyle(UIContextMode mode, Texture background, UIFont font, float fontSize, int width, int buttonHeight) {
+    public UIContextMenuStyle(UIContextMode mode, Texture background, UIFont font, int fontSize, int width, int buttonHeight) {
         this.mode = mode;
         this.background = background;
         this.font = font;
@@ -33,7 +33,7 @@ public class UIContextMenuStyle {
         return font;
     }
 
-    public float getFontSize() {
+    public int getFontSize() {
         return fontSize;
     }
